requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.6b0
cfgv==3.3.0
click==8.0.1
coverage==5.5
distlib==0.3.2
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.8.2
filelock==3.0.12
flake8==3.9.2
graphviz==0.16
gunicorn==20.1.0
identify==2.2.11
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
platformdirs==2.0.2
pluggy==0.13.1
pre-commit==2.13.0
py==1.10.0
pycodestyle==2.7.0
pydotplus==2.0.2
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
virtualenv==20.6.0
